Fix : Debugging components
