+++ src/app/components/home/HeroSection.tsx
@@ -1,0 +1,12 @@
+export default function HeroSection() {
+    return (
+      <section className="relative w-full">
+        <img
+          src="/images/Gambar.jpg"
+          alt="Hero Section"
+          className="w-full h-[400px] object-cover object-[50%_20%]"
+        />
+      </section>
+    );
+  }
+
